==> scripts/R/05_DEGs_sex.Rmd <==
---
title: "Differential expression in each sex"
author: "Kimberly Olney, Ph.D."
date: "09/30/2022"
output:
  html_document:
    df_print: paged
  pdf_document: default
params:
  args: myarg
---
Volcano plots and DEG tables for pairwise comparisons within each sex and sex differential expression. 
# Setup
```{r setup}
knitr::opts_knit$set(root.dir = ".")
```
# User defined variables
```{r set_variables}
source(here::here("scripts/R", "file_paths_and_colours.R"))
source(here::here("scripts/R", "gtf_path.R"))
condition <- c("TYPE.SEX")
tool <- "star"
dge.filtered.norm <- readRDS(paste0("../../rObjects/dge.filtered.norm.rds"))

# add a new column with TYPE + sex inferred
dge.filtered.norm$samples$TYPE_sex_inferred <-
  paste0(dge.filtered.norm$samples$TYPE,
         "_",
         dge.filtered.norm$samples$sex_inferred)
dge.filtered.norm$samples$TYPE_sex_inferred <-
  factor(
    dge.filtered.norm$samples$TYPE_sex_inferred,
    levels = c(
      "CONTROL_male",
      "CONTROL_female",
      "AD_male",
      "AD_female",
      "PA_male",
      "PA_female",
      "LBD_male",
      "LBD_female"
    )
  )
```
# Design
```{r}
design <-
  model.matrix(~ 0 + 
      TYPE_sex_inferred,
    dge.filtered.norm$samples
  )

colnames(design) <-
  c(
    "CONTROL_male", 
    "CONTROL_female", 
    "AD_male", 
    "AD_female", 
    "PA_male",
    "PA_female", 
    "LBD_male",
    "LBD_female"
  )
```
# Contrast
```{r}
contrasts <- makeContrasts(
  LBDvsControl_no_sex_correction = (LBD_male + LBD_female)/2 - (CONTROL_male + CONTROL_female)/2,
  LBDvsAD_no_sex_correction  = (LBD_male + LBD_female)/2 - (AD_male + AD_female)/2,
  LBDvsPA_no_sex_correction  = (LBD_male + LBD_female)/2 - (PA_male + PA_female)/2,
  ADvsControl_no_sex_correction  = (AD_male + AD_female)/2 - (CONTROL_male + CONTROL_female)/2,
  PAvsControl_no_sex_correction  = (PA_male + PA_female)/2 - (CONTROL_male + CONTROL_female)/2,
  ADvsPA_no_sex_correction  = (AD_male + AD_female)/2 - (PA_male + PA_female)/2,
  male_XY_LBDvsControl = LBD_male - CONTROL_male,
  male_XY_LBDvsAD = LBD_male  - AD_male,
  male_XY_LBDvsPA = LBD_male - PA_male,
  male_XY_ADvsControl = AD_male - CONTROL_male,
  male_XY_PAvsControl = PA_male - CONTROL_male,
  male_XY_ADvsPA = AD_male  - PA_male,
  female_XX_LBDvsControl = LBD_female - CONTROL_female,
  female_XX_LBDvsAD = LBD_female  - AD_female,
  female_XX_LBDvsPA = LBD_female - PA_female,
  female_XX_ADvsControl = AD_female - CONTROL_female,
  female_XX_PAvsControl = PA_female - CONTROL_female,
  female_XX_ADvsPA = AD_female  - PA_female,
  female_LBD_vs_male_LBD = LBD_female - LBD_male, 
  female_AD_vs_male_AD = AD_female - AD_male, 
  female_PA_vs_male_PA = PA_female - PA_male, 
  female_Control_vs_male_Control = CONTROL_female - CONTROL_male, 
  ADvsControl_sex_interaction = ((AD_female -  CONTROL_female) - (AD_male - CONTROL_male)),
  PAvsControl_sex_interaction = ((PA_female - CONTROL_female) - (PA_male - CONTROL_male)),
  LBDvsControl_sex_interaction = ((LBD_female - CONTROL_female) - (LBD_male - CONTROL_male)),
  LBDvsAD_sex_interaction = ((LBD_female - AD_female ) - (LBD_male - AD_male)),
  LBDvsPA_sex_interaction = ((LBD_female - PA_female ) - (LBD_male - PA_male)),
  ADvsPA_sex_interaction = ((AD_female - PA_female) - (AD_male - PA_male)),
  levels = colnames(design))
contrasts
allComparisons <- colnames(contrasts)
```

# Volcano plots
```{r volcano, warning=FALSE}
for (i in allComparisons) {
  group1_vs_group2 <-
    read.delim(
      paste0(
        "../../results/",
        tool,
        "/DEGs/",
        condition,
        "_",
        i,
        "_gene_DEGs_FDRq1.00.txt"
      )
    )
  color_values <- vector()
  max <- nrow(group1_vs_group2)
  for (row in 1:max) {
    if (group1_vs_group2$adj.P.Val[row] < 0.05) {
      if (group1_vs_group2$logFC [row] > 0.25) {
        color_values <- c(color_values, 1)
      }
      else if (group1_vs_group2$logFC[row] < -.25) {
        color_values <- c(color_values, 2)
      }
      else {
        color_values <- c(color_values, 3)
      }
    }
    else{
      color_values <- c(color_values, 3)
    }
  }
  group1_vs_group2$color_adjpval_0.05 <- factor(color_values)
  data <- group1_vs_group2
  # plot only if there are DEGs with p_val_adj < 0.05
  num <- subset(data, (adj.P.Val < 0.05 & logFC < -.25)  | (adj.P.Val < 0.05 & logFC > .25 ))
  num <- nrow(num)
  if (num != 0) {
    up <- data[data$color_adjpval_0.05 == 1,]
    up10 <- up[1:10,]
    upFold <- subset(up, logFC > 1)
    upFold <- upFold[!(upFold$gene_name %in% up10$gene_name),]
    down <- data[data$color_adjpval_0.05 == 2,]
    down10 <- down[1:10,]
    downFold <- subset(down, logFC < -1)
    downFold <- downFold[!(downFold$gene_name %in% down10$gene_name),]
    if (!1 %in% unique(data$color_adjpval_0.05)) {
      my_colors <- c("blue", "gray")
    } else if (!2 %in% unique(data$color_adjpval_0.05)) {
      my_colors <- c("red", "gray")
    } else if (!1 %in% unique(data$color_adjpval_0.05) &&
               !2 %in% unique(data$color_adjpval_0.05)) {
      my_colors <- c("gray")
    } else {
      my_colors <- c("red", "blue", "gray")
    }
    hadjpval <- (-log10(max(data$P.Value[data$adj.P.Val < 0.05],
                            na.rm = TRUE)))
    negFC <- c(-.25)
    posFC <- c(.25) 
    i <- gsub("_", " ", i)
    p <-
      ggplot(data = data,
             aes(
               x = logFC,
               y = -log10(P.Value),
               color = color_adjpval_0.05
             )) +
      geom_point(alpha = 0.8, size = 1) +
      theme_bw() +
      theme(legend.position = "none") +
      scale_color_manual(values = my_colors) +
      labs(
        title = "",
        x = expression(log[2](FC)),
        y = expression(-log[10] ~ "(" ~ italic("p") ~ "-value)")
      )  +
      geom_hline(yintercept = hadjpval,
                 #  horizontal line
                 colour = "#000000",
                 linetype = "dashed") +
      geom_vline(xintercept = negFC,
                 #  horizontal line
                 colour = "#000000",
                 linetype = "dashed") +
      geom_vline(xintercept = posFC,
                 #  horizontal line
                 colour = "#000000",
                 linetype = "dashed") +
      ggtitle(paste0(i)) +
      geom_text_repel(
        data = up10,
        aes(
          x = logFC,
          y = -log10(P.Value),
          label = gene_name
        ),
        color = "maroon",
        fontface = "italic",
        size = 3,
        max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
      ) +
      geom_text_repel(
        data = upFold,
        aes(
          x = logFC,
          y = -log10(P.Value),
          label = gene_name
        ),
        color = "maroon",
        fontface = "italic",
        size = 3,
        max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
      ) +
      geom_text_repel(
        data = down10,
        aes(
          x = logFC,
          y = -log10(P.Value),
          label = gene_name
        ),
        color = "navyblue",
        fontface = "italic",
        size = 3,
        max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
      ) +
      geom_text_repel(
        data = downFold,
        aes(
          x = logFC,
          y = -log10(P.Value),
          label = gene_name
        ),
        color = "navyblue",
        fontface = "italic",
        size = 3,
        max.overlaps = getOption("ggrepel.max.overlaps", default = 10)
      )
    p
    i <- gsub(" ", "_", i)
    # save
    path <-
      paste0(
        "../../results/",
        tool,
        "/volcano/",
        condition,
        "_",
        i,
        "_gene_volcano_FDRq0.05"
      )
    pdf(paste0(path, ".pdf"),
        height = 5,
        width = 5)
    print(p)
    dev.off()
    
    print(paste("i =", i))
  }
} 
```

```{r cleanup}
# clean up
remove(up, up10, upFold, group1_vs_group2, downFold, down10, data, p)
```
# Comparisons list 
```{r comparison_lists, eval=FALSE}
# within each sex 
WithinEachSexComparisons <- c("female_XX_PAvsControl", "female_XX_ADvsControl", "female_XX_LBDvsControl", "female_XX_ADvsPA", "female_XX_LBDvsPA", "female_XX_LBDvsAD", "male_XY_PAvsControl", "male_XY_ADvsControl", "male_XY_LBDvsControl", "male_XY_ADvsPA", "male_XY_LBDvsPA", "male_XY_LBDvsAD")

# FvsM within each TYPE
FvsM_Comparisons <- c("female_Control_vs_male_Control", "female_PA_vs_male_PA", "female_AD_vs_male_AD", "female_LBD_vs_male_LBD")

# sex interaction
SexInteraction <- c("PAvsControl_sex_interaction","ADvsControl_sex_interaction","LBDvsControl_sex_interaction","ADvsPA_sex_interaction","ADvsPA_sex_interaction","LBDvsAD_sex_interaction")
```

# Make excel table
```{r excel}
# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/DEGs/", condition, "_", i,
    "_gene_DEGs_FDRq1.00.txt"
  )
  assign(paste0(i),
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ))
}
```
make list for each comparisons groups 
each new list will be separate excel table 
```{r}
list_of_datasets <- list(
"female_XX_PAvsControl" = female_XX_PAvsControl,
"female_XX_ADvsControl" = female_XX_ADvsControl,
"female_XX_LBDvsControl" = female_XX_LBDvsControl,
"female_XX_ADvsPA" = female_XX_ADvsPA,
"female_XX_LBDvsPA" =  female_XX_LBDvsPA, 
"female_XX_LBDvsAD" = female_XX_LBDvsAD,
"male_XY_PAvsControl" = male_XY_PAvsControl,
"male_XY_ADvsControl" = male_XY_ADvsControl,
"male_XY_LBDvsControl" = male_XY_LBDvsControl,
"male_XY_ADvsPA" = male_XY_ADvsPA,
"male_XY_LBDvsPA" =  male_XY_LBDvsPA, 
"male_XY_LBDvsAD" = male_XY_LBDvsAD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_WithinEachSexComparisons.DEGs.FDRq1.00.xlsx"))
remove(list_of_datasets)


list_of_datasets <- list(
"FvsM_Control" = female_Control_vs_male_Control,
"FvsM_PA" = female_PA_vs_male_PA,
"FvsM_AD" = female_AD_vs_male_AD,
"FvsM_LBD" = female_LBD_vs_male_LBD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_FvsM_Comparisons.DEGs.FDRq1.00.xlsx"))
remove(list_of_datasets)


list_of_datasets <- list(
"FvsM_Control" = female_Control_vs_male_Control,
"FvsM_PA" = female_PA_vs_male_PA,
"FvsM_AD" = female_AD_vs_male_AD,
"FvsM_LBD" = female_LBD_vs_male_LBD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_FvsM_Comparisons.DEGs.FDRq1.00.xlsx"))
remove(list_of_datasets)


list_of_datasets <- list(
"PAvsControl_sex_interaction" = PAvsControl_sex_interaction,
"ADvsControl_sex_interaction" = ADvsControl_sex_interaction,
"LBDvsControl_sex_interaction" = LBDvsControl_sex_interaction,
"ADvsPA_sex_interaction" = ADvsPA_sex_interaction,
"LBDvsPA_sex_interaction" = LBDvsPA_sex_interaction,
"LBDvsAD_sex_interaction" = LBDvsAD_sex_interaction)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_SexInteraction.DEGs.FDRq1.00.xlsx"))
remove(list_of_datasets)
```
clean up
```{r cleanup_excel}
remove(
female_XX_PAvsControl,
female_XX_ADvsControl,
female_XX_LBDvsControl,
female_XX_ADvsPA,
female_XX_LBDvsPA,
female_XX_LBDvsAD,
male_XY_PAvsControl,
male_XY_ADvsControl,
male_XY_LBDvsControl,
male_XY_ADvsPA,
male_XY_LBDvsPA,
male_XY_LBDvsAD,
female_Control_vs_male_Control,
female_PA_vs_male_PA,
female_AD_vs_male_AD,
female_LBD_vs_male_LBD,
PAvsControl_sex_interaction,
ADvsControl_sex_interaction,
LBDvsControl_sex_interaction,
ADvsPA_sex_interaction,
LBDvsPA_sex_interaction,
LBDvsAD_sex_interaction)
```
# Excel table of significant DEGs
```{r}
# first read in the files
for (i in allComparisons) {
  filepath <- paste0("../../results/",tool,
    "/DEGs/", condition, "_", i,
    "_gene_DEGs_FDRq0.05_logFC_0.25.txt"
  )
  assign(paste0(i),
         tryCatch(
         read.delim(
           filepath,
           header = TRUE,
           sep = "\t",
           stringsAsFactors = FALSE
         ), error=function(e) NULL))
}
list_of_datasets <- list(
"female_XX_PAvsControl" = female_XX_PAvsControl,
"female_XX_ADvsControl" = female_XX_ADvsControl,
"female_XX_LBDvsControl" = female_XX_LBDvsControl,
"female_XX_ADvsPA" = female_XX_ADvsPA,
"female_XX_LBDvsPA" =  female_XX_LBDvsPA, 
"female_XX_LBDvsAD" = female_XX_LBDvsAD,
"male_XY_PAvsControl" = male_XY_PAvsControl,
"male_XY_ADvsControl" = male_XY_ADvsControl,
"male_XY_LBDvsControl" = male_XY_LBDvsControl,
"male_XY_ADvsPA" = male_XY_ADvsPA,
"male_XY_LBDvsPA" =  male_XY_LBDvsPA, 
"male_XY_LBDvsAD" = male_XY_LBDvsAD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_WithinEachSexComparisons.DEGs.FDRq0.05_logFC_0.25.xlsx"))
remove(list_of_datasets)

list_of_datasets <- list(
"FvsM_Control" = female_Control_vs_male_Control,
"FvsM_PA" = female_PA_vs_male_PA,
"FvsM_AD" = female_AD_vs_male_AD,
"FvsM_LBD" = female_LBD_vs_male_LBD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_FvsM_Comparisons.DEGs.FDRq0.05_logFC_0.25.xlsx"))
remove(list_of_datasets)

list_of_datasets <- list(
"FvsM_Control" = female_Control_vs_male_Control,
"FvsM_PA" = female_PA_vs_male_PA,
"FvsM_AD" = female_AD_vs_male_AD,
"FvsM_LBD" = female_LBD_vs_male_LBD)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_FvsM_Comparisons.DEGs.FDRq0.05_logFC_0.25.xlsx"))
remove(list_of_datasets)

list_of_datasets <- list(
"PAvsControl_sex_interaction" = PAvsControl_sex_interaction,
"ADvsControl_sex_interaction" = ADvsControl_sex_interaction,
"LBDvsControl_sex_interaction" = LBDvsControl_sex_interaction,
"ADvsPA_sex_interaction" = ADvsPA_sex_interaction,
"LBDvsPA_sex_interaction" = LBDvsPA_sex_interaction,
"LBDvsAD_sex_interaction" = LBDvsAD_sex_interaction)
write.xlsx(list_of_datasets, file = paste0(
    "../../results/",
    tool,
    "/DEGs/", condition,"_SexInteraction.DEGs.FDRq0.05_logFC_0.25.xlsx"))
remove(list_of_datasets)
```
```{r}
```
